Add backwards compatibility embed code for older presentations.
Older presentations used the following URL format:

     http://docs.google.com/present/view?id=xxxx

Previously, the plugin did not support presentation URLs in that format.
Now it does.

----

This commit also adds support for the start parameter for Presentations.

If you do this:

     [gdoc link=https://docs.google.com/presentation/pub?id=XXX start=1]

This will autostart the presentation after the entire page loads.

Note: This parameter will *not* work for older presentations.