Add 'type' shortcode attribute.
For files that are not a native Google Doc format, add a 'type' attribute
so users can set the type in the shortcode.

This commit adds initial support for audio files (type='audio').  Tested
with MP3s so far.  Link must be generated using the 'Anyone can view the
link' format.

See #7.